+++ lib/HomePage/HomePage.dart
@@ -1,11 +1,9 @@
-import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
-import '../SchedulePage/LegendItem.dart';
-import 'DailyProgress.dart';
 import '../Utils/FetchRecord.dart';
-import 'MonthlyCalender.dart';
-import 'SubjectDistributionPlot.dart';
-import 'WeeklyProgress.dart';
+import 'DailyProgressCard.dart';
+import 'ProgressCalendarCard.dart';
+import 'SubjectDistributionCard.dart';
+import 'WeeklyProgressCard.dart';
 import 'calculation_utils.dart';
 
 class HomePage extends StatefulWidget {
@@ -315,9 +313,9 @@ class _HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin 
           flex: 1,
           child: Column(
             children: [
-              _buildDailyProgressCard(allRecords, cardPadding),
+              buildDailyProgressCard(allRecords, cardPadding,context),
               const SizedBox(height: 32),
-              _buildSubjectDistributionCard(subjectDistribution, cardPadding),
+              buildSubjectDistributionCard(subjectDistribution, cardPadding,context),
             ],
           ),
         ),
@@ -327,9 +325,9 @@ class _HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin 
           flex: 1,
           child: Column(
             children: [
-              _buildWeeklyProgressCard(allRecords, cardPadding),
+              buildWeeklyProgressCard(allRecords, cardPadding,context),
               const SizedBox(height: 32),
-              _buildProgressCalendarCard(allRecords, cardPadding),
+              buildProgressCalendarCard(allRecords, cardPadding,context),
             ],
           ),
         ),
@@ -344,295 +342,17 @@ class _HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin 
       double cardPadding) {
     return Column(
       children: [
-        _buildDailyProgressCard(allRecords, cardPadding),
+        buildDailyProgressCard(allRecords, cardPadding,context),
         const SizedBox(height: 24),
-        _buildWeeklyProgressCard(allRecords, cardPadding),
+        buildWeeklyProgressCard(allRecords, cardPadding,context),
         const SizedBox(height: 24),
-        _buildProgressCalendarCard(allRecords, cardPadding),
+        buildProgressCalendarCard(allRecords, cardPadding,context),
         const SizedBox(height: 24),
-        _buildSubjectDistributionCard(subjectDistribution, cardPadding),
+        buildSubjectDistributionCard(subjectDistribution, cardPadding,context),
       ],
     );
   }
 
-  Widget _buildDailyProgressCard(List<Map<String, dynamic>> allRecords, double cardPadding) {
-    return Container(
-      width: double.infinity,
-      decoration: BoxDecoration(
-        color: Theme.of(context).cardColor,
-        borderRadius: BorderRadius.circular(24),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.05),
-            blurRadius: 20,
-            offset: const Offset(0, 4),
-          ),
-        ],
-      ),
-      padding: EdgeInsets.all(cardPadding),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            'Daily Progress',
-            style: TextStyle(
-              fontSize: 18,
-              fontWeight: FontWeight.w600,
-              color: Theme.of(context).textTheme.titleLarge?.color,
-            ),
-          ),
-          const SizedBox(height: 24),
-          SizedBox(
-            height: 250,
-            width: double.infinity,
-            child: LineChart(createLineChartData(allRecords)),
-          ),
-          const SizedBox(height: 24),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: const [
-              LegendItem(label: 'Lectures', color: Colors.blue, icon: Icons.school),
-              LegendItem(label: 'Revisions', color: Colors.green, icon: Icons.check_circle),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildWeeklyProgressCard(List<Map<String, dynamic>> allRecords, double cardPadding) {
-    return Container(
-      width: double.infinity,
-      decoration: BoxDecoration(
-        color: Theme.of(context).cardColor,
-        borderRadius: BorderRadius.circular(24),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.05),
-            blurRadius: 20,
-            offset: const Offset(0, 4),
-          ),
-        ],
-      ),
-      padding: EdgeInsets.all(cardPadding),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            'Weekly Progress',
-            style: TextStyle(
-              fontSize: 18,
-              fontWeight: FontWeight.w600,
-              color: Theme.of(context).textTheme.titleLarge?.color,
-            ),
-          ),
-          const SizedBox(height: 24),
-          SizedBox(
-            height: 250,
-            width: double.infinity,
-            child: BarChart(createBarChartWeeklyData(allRecords)),
-          ),
-          const SizedBox(height: 16),
-          _buildLegend(),
-          const SizedBox(height: 8),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildLegend() {
-    return const Row(
-      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      children: [
-        LegendItem(label: 'Lectures', color: Colors.blue, icon: Icons.school),
-        LegendItem(label: 'Revisions', color: Colors.green, icon: Icons.check_circle),
-        LegendItem(label: 'Missed', color: Colors.red, icon: Icons.cancel),
-        LegendItem(label: 'Scheduled', color: Colors.orange, icon: Icons.schedule),
-      ],
-    );
-  }
-
-  Widget _buildProgressCalendarCard(List<Map<String, dynamic>> allRecords, double cardPadding) {
-    return Container(
-      width: double.infinity,
-      decoration: BoxDecoration(
-        color: Theme.of(context).cardColor,
-        borderRadius: BorderRadius.circular(24),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.05),
-            blurRadius: 20,
-            offset: const Offset(0, 4),
-          ),
-        ],
-      ),
-      padding: EdgeInsets.all(cardPadding),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            'Progress Calendar',
-            style: TextStyle(
-              fontSize: 18,
-              fontWeight: FontWeight.w600,
-              color: Theme.of(context).textTheme.titleLarge?.color,
-            ),
-          ),
-          const SizedBox(height: 24),
-          SizedBox(
-            height: 750,
-            child: StudyCalendar(
-              records: allRecords,
-            ),
-          ),
-          const SizedBox(height: 16),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSubjectDistributionCard(Map<String, int> subjectDistribution, double cardPadding) {
-    // Get the screen width to calculate responsive sizes
-    final screenWidth = MediaQuery.of(context).size.width;
-
-    // Calculate dynamic radius based on screen width
-    final bool isSmallScreen = screenWidth < 600;
-    final double chartRadius = isSmallScreen ? 80 : 100;
-    final double centerRadius = isSmallScreen ? 60 : 80;
-
-    return Container(
-      width: double.infinity,
-      decoration: BoxDecoration(
-        color: Theme.of(context).cardColor,
-        borderRadius: BorderRadius.circular(24),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.05),
-            blurRadius: 20,
-            offset: const Offset(0, 4),
-          ),
-        ],
-      ),
-      padding: EdgeInsets.all(cardPadding),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            'Subject Distribution',
-            style: TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.w600,
-              color: Theme.of(context).textTheme.titleLarge?.color,
-            ),
-          ),
-          const SizedBox(height: 24),
-          // Use LayoutBuilder to make pie chart responsive to its container
-          LayoutBuilder(
-              builder: (context, constraints) {
-                return SizedBox(
-                  height: 500,
-                  child: PieChart(
-                    PieChartData(
-                      sections: createPieChartSections(
-                        subjectDistribution,
-                        chartRadius,
-                        Theme.of(context),
-                      ),
-                      sectionsSpace: 2,
-                      centerSpaceRadius: centerRadius,
-                      borderData: FlBorderData(show: false),
-                      pieTouchData: PieTouchData(
-                        touchCallback: (FlTouchEvent event, pieTouchResponse) {
-                          // Could implement hover effects or selection here
-                        },
-                        enabled: true,
-                      ),
-                    ),
-                  ),
-                );
-              }
-          ),
-          const SizedBox(height: 32),
-          Center(
-            child: buildPieChartLegend(subjectDistribution, context),
-          ),
-        ],
-      ),
-    );
-  }
-
-
-  Widget buildPieChartLegend(Map<String, int> subjectCounts, BuildContext context) {
-    // Modern color palette - same as in createPieChartSections
-    final colors = [
-      const Color(0xFF5038BC),  // Deep purple
-      const Color(0xFF4ECDC4),  // Teal
-      const Color(0xFFFF6B6B),  // Coral
-      const Color(0xFFFFD166),  // Yellow
-      const Color(0xFF118AB2),  // Blue
-      const Color(0xFFEF8354),  // Orange
-      const Color(0xFF06D6A0),  // Mint
-      const Color(0xFFDA627D),  // Pink
-    ];
-
-    // Calculate total count for percentage calculation
-    int totalCount = subjectCounts.values.fold(0, (sum, count) => sum + count);
-
-    // Sort entries by count for better visualization (matching the pie chart order)
-    var sortedEntries = subjectCounts.entries.toList()
-      ..sort((a, b) => b.value.compareTo(a.value));
-
-    return Container(
-      padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
-      decoration: BoxDecoration(
-        color: Theme.of(context).cardColor,
-        borderRadius: BorderRadius.circular(16),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.03),
-            blurRadius: 8,
-            offset: const Offset(0, 2),
-          ),
-        ],
-      ),
-      child: Wrap(
-        spacing: 16,
-        runSpacing: 16,
-        alignment: WrapAlignment.center,
-        children: sortedEntries.asMap().entries.map((mapEntry) {
-          int index = mapEntry.key;
-          var entry = mapEntry.value;
-          double percentage = totalCount > 0 ? (entry.value / totalCount * 100) : 0;
-
-          return Container(
-            margin: const EdgeInsets.symmetric(horizontal: 4),
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Container(
-                  width: 14,
-                  height: 14,
-                  decoration: BoxDecoration(
-                    color: colors[index % colors.length],
-                    borderRadius: BorderRadius.circular(4),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Text(
-                  '${entry.key} (${percentage.toStringAsFixed(1)}%)',
-                  style: TextStyle(
-                    fontSize: 14,
-                    color: Theme.of(context).textTheme.bodyLarge?.color,
-                    fontWeight: FontWeight.w500,
-                  ),
-                ),
-              ],
-            ),
-          );
-        }).toList(),
-      ),
-    );
-  }
 
   Widget _buildStatCard(String title, String value, Color color) {
     return Container(
